+++ lib/modules/security/security_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: security
-// Updated at: 2026-05-14 12:32:39
-// Commit: feat(security): implement shimmer loading skeleton in security
+// Updated at: 2026-05-14 14:42:56
+// Commit: feat(security): add bi-directional event bus handler for security
 
 class securityComponentState {
-  final String commitRef = '1136';
-  final String lastMessage = 'feat(security): implement shimmer loading skeleton in security';
-  final DateTime timestamp = DateTime.parse('2026-05-14T12:32:39+05:00');
+  final String commitRef = '1143';
+  final String lastMessage = 'feat(security): add bi-directional event bus handler for security';
+  final DateTime timestamp = DateTime.parse('2026-05-14T14:42:56+05:00');
 }